Add files via upload
##Code Problem Statement - 

Create a random guessing game for a player to guess a correct number between the range of 1 and 11. The player will have 3 (three) attempts at guessing the correct answer.

After each attempt, the user sees the results of the guess made, either correct or wrong and game terminates if the user gets the answer in less than 3 tries.
In this case, the answer is notified that the guess was correct and will be requested to either continue the or end the game

In the case that the player fails at the three attempts, the game will ask if the user wants to continue the game or have a rerun.